Remove deprecated image inputs, rename where appropriate
