+++ app/javascript/application.js
@@ -4,5 +4,5 @@ import "controllers";
 
 // Custom code
 // Be sure to add to config/importmap.rb and restart the server to see the changes
-import "avatar_input_preview";
 import "filter_form_tracker";
+import "image_input_preview";
