+++ config/importmap.rb
@@ -6,5 +6,5 @@ pin "@hotwired/stimulus", to: "stimulus.min.js"
 pin "@hotwired/stimulus-loading", to: "stimulus-loading.js"
 pin_all_from "app/javascript/controllers", under: "controllers"
 
-pin "avatar_input_preview"
 pin "filter_form_tracker"
+pin "image_input_preview"
